+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 17/03/2012
 - Quitar las votaciones
+- Crear un entorno staging para heroku
 
 12/03/2012
 - Añadir tests